feat: login and logout

+++ pure_plate/account/views.py
@@ -2,28 +2,50 @@ from rest_framework import generics, permissions
 from rest_framework.response import Response
 from rest_framework.authtoken.models import Token
 from rest_framework.views import APIView
-from django.contrib.auth import authenticate
 from .models import User
-from .serializers import UserSerializer
+from rest_framework import status
+from rest_framework.response import Response
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import AllowAny
+from .serializers import UserRegisterSerializer, UserLoginSerializer
+from .models import User
 
-class RegisterAPIView(generics.CreateAPIView):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-    permission_classes = []  # 인증이 필요하지 않도록 설정
+@api_view(['POST'])
+@permission_classes([AllowAny])
+def register(request):
+    if request.method == 'POST':
+        serializer = UserRegisterSerializer(data=request.data)
+        if not serializer.is_valid(raise_exception=True):
+            return Response({"message": "Request Body Error."}, status=status.HTTP_409_CONFLICT)
 
-class LoginAPIView(APIView):
-    def post(self, request):
-        email = request.data.get('email')
-        password = request.data.get('password')
-        user = authenticate(email=email, password=password)
-        if user:
-            token, created = Token.objects.get_or_create(user=user)
-            return Response({'token': token.key})
-        return Response({'error': 'Invalid Credentials'}, status=400)
+        if User.objects.filter(email=serializer.validated_data['email']).first() is None:
+            serializer.save()
+            return Response({"message": "ok"}, status=status.HTTP_201_CREATED)
+        return Response({"message": "duplicate email"}, status=status.HTTP_409_CONFLICT)
 
-class LogoutAPIView(APIView):
-    permission_classes = (permissions.IsAuthenticated,)
 
-    def post(self, request):
+@api_view(['POST'])
+@permission_classes([AllowAny])
+def login(request):
+    if request.method == 'POST':
+        serializer = UserLoginSerializer(data=request.data)
+        if not serializer.is_valid(raise_exception=True):
+            return Response({"message": "Request Body Error."}, status=status.HTTP_409_CONFLICT)
+        if serializer.validated_data['email'] == "None":
+            return Response({"message": "fail"}, status=status.HTTP_401_UNAUTHORIZED)
+        
+        response = {
+            'success': 'True'
+        }
+        return Response(response, status=status.HTTP_200_OK)
+    
+
+
+@api_view(['POST'])
+@permission_classes([permissions.IsAuthenticated])
+def logout(request):
+    if request.method == 'POST':
+        # 사용자의 토큰을 삭제하여 로그아웃 처리
         request.user.auth_token.delete()
-        return Response(status=204)
+        return Response({"message": "Logged out successfully."}, status=status.HTTP_200_OK)
+
